Creating authorization on the index view
### What's done:
- Polishing frontend
- Adding Auth to the index view
