+++ save-deploy/README.md
@@ -92,9 +92,8 @@ Usually, not the whole stack is required for development. Application logic is p
   Run `./gradlew -Psave.profile=dev :save-frontend:run` to start save-frontend using webpack-dev-server, requests to REST API will be
   proxied as configured in [dev-server.js](../save-frontend/webpack.config.d/dev-server.js).
 * For developing most part of platform's logic, the above will be enough. If local testing of authentication flow is required, however,
-  `api-gateway` can be run locally together with [dex](https://github.com/dexidp/dex) OAuth2 server. In order to do so, run 
-  `docker compose up -d dex` and then start `api-gateway` with `dev` profile enabled. Using [`application-dev.yaml`](../api-gateway/src/main/resources/application-dev.yml)
-  one can connect gateway with dev build of frontend running with webpack by changing `gateway.frontend.url`.
+  `api-gateway` can be run locally together with [dex](https://github.com/dexidp/dex) OAuth2 server. [`application-dev.yaml`](../api-gateway/src/main/resources/application-dev.yml)
+  already contains configuration for dev authorization providers (see the description below).
 
 ### Using OAuth with a local deployment
 
@@ -103,7 +102,7 @@ Usually, not the whole stack is required for development. Application logic is p
    omitting any gateway. When enabling OAuth, make sure the gateway is contacted
    instead:
  
-   * `context`: add `/sec/**`, `/oauth2/**`, and `/login/oauth2/**` to the list;
+   * `context`: add `/sec/**, /oauth2/**, /login/oauth2/**` to the list;
    * `target`: change to [`http://localhost:5300`](http://localhost:5300) (the
      default gateway URL); 
    * `onProxyReq`: drop the entire callback, since both headers (`Authorization`
