+++ save-frontend/src/main/kotlin/com/saveourtool/save/frontend/components/views/fossgraph/FossGraphView.kt
@@ -39,6 +39,8 @@ import kotlinx.serialization.json.Json
     "TYPE_ALIAS",
 )
 val fossGraph: FC<FossGraphViewProps> = FC { props ->
+
+    particles()
     useBackground(Style.WHITE)
     useTooltip()
